requirements.txt
streamlit
yfinance
requests
google-generativeai
python-dotenv
pandas
beautifulsoup4
lxml
plotly
openpyxl  
WeasyPrint